implement strictly positional things
