+++ src/item.rs
@@ -5,6 +5,7 @@ use crate::{Meta, Named};
 pub enum Item {
     Positional {
         metavar: &'static str,
+        strict: bool,
         help: Option<String>,
     },
     Command {
@@ -50,7 +51,11 @@ impl From<&Named> for ShortLong {
 impl std::fmt::Display for Item {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         match self {
-            Item::Positional { metavar, help: _ } => write!(f, "<{}>", metavar),
+            Item::Positional {
+                metavar,
+                help: _,
+                strict: _,
+            } => write!(f, "<{}>", metavar),
             Item::Command { .. } => write!(f, "COMMAND ..."),
             Item::Flag { name, help: _ } => write!(f, "{}", name),
             Item::Argument {
