+++ src/meta_usage.rs
@@ -15,6 +15,7 @@ pub(crate) enum UsageMeta {
     LongFlag(&'static str),
     LongArg(&'static str, &'static str),
     Pos(&'static str),
+    StrictPos(&'static str),
     Command,
 }
 
@@ -22,6 +23,12 @@ pub(crate) enum UsageMeta {
 // - any meta containing a positional becomes positional itself
 // - positional item must occupy the right most position inside Meta::And
 
+pub(crate) fn to_usage_meta(meta: &Meta) -> Option<UsageMeta> {
+    let mut had_commands = false;
+    let mut is_pos = false;
+
+    collect_usage_meta(meta, false, &mut had_commands, &mut is_pos)
+}
 /// Transforms `Meta` to [`UsageMeta`]
 ///
 /// parameter `required` defines the value's context: optional or required.
@@ -31,7 +38,7 @@ pub(crate) enum UsageMeta {
 /// `bpaf` uses parameter `is_pos` for positional validation, initialize it with false
 ///
 /// return value is None if parser takes no parameters at all
-pub(crate) fn collect_usage_meta(
+fn collect_usage_meta(
     meta: &Meta,
     required: bool,
     had_commands: &mut bool,
@@ -94,9 +101,17 @@ pub(crate) fn collect_usage_meta(
         Meta::Decorated(meta, _) => collect_usage_meta(meta, required, had_commands, is_pos)?,
         Meta::Skip => return None,
         Meta::Item(i) => match i {
-            Item::Positional { metavar, help: _ } => {
+            Item::Positional {
+                metavar,
+                strict,
+                help: _,
+            } => {
                 *is_pos = true;
-                UsageMeta::Pos(metavar)
+                if *strict {
+                    UsageMeta::StrictPos(metavar)
+                } else {
+                    UsageMeta::Pos(metavar)
+                }
             }
             Item::Command {
                 name: _,
@@ -179,6 +194,7 @@ impl std::fmt::Display for UsageMeta {
             UsageMeta::LongArg(l, v) => write!(f, "--{} {}", l, v),
             UsageMeta::Command => f.write_str("COMMAND ..."),
             UsageMeta::Pos(s) => write!(f, "<{}>", s),
+            UsageMeta::StrictPos(s) => write!(f, "-- <{}>", s),
         }
     }
 }
