+++ src/tests.rs
@@ -1,7 +1,7 @@
 #[test]
 #[cfg(any(windows, unix))]
 fn wtf_shenanigans_1() {
-    use crate::args::{split_os_argument, word, ArgType};
+    use crate::args::{split_os_argument, word_arg, ArgType};
     use std::ffi::OsString;
 
     for (i_c, prefix) in [
@@ -34,13 +34,13 @@ fn wtf_shenanigans_1() {
         let (o_c, o_prefix, o_suffix) = split_os_argument(&os_string).unwrap();
         assert_eq!(i_c, o_c);
         assert_eq!(i_prefix.to_str().unwrap(), o_prefix);
-        assert_eq!(word(i_suffix), o_suffix.unwrap());
+        assert_eq!(word_arg(i_suffix, false), o_suffix.unwrap());
     }
 }
 
 #[test]
 fn wtf_shenanigans_2() {
-    use crate::args::{split_os_argument, split_os_argument_fallback, word, ArgType};
+    use crate::args::{split_os_argument, split_os_argument_fallback, word_arg, ArgType};
     use std::ffi::OsString;
 
     for (i_c, prefix) in [
@@ -63,16 +63,11 @@ fn wtf_shenanigans_2() {
         let (o_c, o_prefix, o_suffix) = split_os_argument(&os_string).unwrap();
         assert_eq!(i_c, o_c);
         assert_eq!(i_prefix.to_str().unwrap(), o_prefix);
-        assert_eq!(word(i_suffix.clone()), o_suffix.unwrap());
+        assert_eq!(word_arg(i_suffix.clone(), false), o_suffix.unwrap());
 
         let (o_c, o_prefix, o_suffix) = split_os_argument_fallback(&os_string).unwrap();
         assert_eq!(i_c, o_c);
         assert_eq!(i_prefix.to_str().unwrap(), o_prefix);
-        assert_eq!(word(i_suffix.clone()), o_suffix.unwrap());
+        assert_eq!(word_arg(i_suffix.clone(), false), o_suffix.unwrap());
     }
-}
-
-#[todo]
-fn asdf() {
-    todo!("extract stuff out of ParseThisOrThat - only parsers need to be monomorphic, everything else just picks what to use");
 }
