+++ tests/tests.rs
@@ -538,24 +538,6 @@ Available options:
     parser.run_inner(Args::from(&["-c"])).unwrap();
 }
 
-#[test]
-fn positional_argument() {
-    let p = positional("FILE").group_help("File to process");
-    let parser = p.to_options();
-
-    let help = parser
-        .run_inner(Args::from(&["--help"]))
-        .unwrap_err()
-        .unwrap_stdout();
-    let expected_help = "\
-Usage: <FILE>
-
-Available options:
-    -h, --help  Prints help information
-";
-    assert_eq!(expected_help, help);
-}
-
 mod git {
     use super::*;
 
@@ -805,19 +787,6 @@ fn long_path_in_construct() {
 }
 
 #[test]
-fn helpful_error_message() {
-    let parser = positional("FOO")
-        .some("You need to specify at least one FOO")
-        .to_options();
-
-    let err = parser
-        .run_inner(Args::from(&[]))
-        .unwrap_err()
-        .unwrap_stderr();
-    assert_eq!("You need to specify at least one FOO", err);
-}
-
-#[test]
 fn env_variable() {
     let name = "BPAF_SECRET_API_KEY";
     let parser = long("key")
@@ -883,13 +852,6 @@ Available options:
     -h, --help  Prints help information
 ";
     assert_eq!(expected_help, help);
-}
-
-#[test]
-fn dash_is_positional() {
-    let a = positional("FILE");
-    let parser = a.to_options();
-    assert_eq!("-", parser.run_inner(Args::from(&["-"])).unwrap());
 }
 
 #[test]
@@ -992,55 +954,6 @@ Available options:
 
     // and "k" isn't a thing
     parser.run_inner(Args::from(&["k"])).unwrap_err();
-}
-
-#[test]
-fn positional_with_help() {
-    let user = positional("USER").help("github user\nin two lines");
-    let api = positional("API_KEY").help("api key to use");
-    let parser = construct!(user, api).to_options();
-
-    let help = parser
-        .run_inner(Args::from(&["--help"]))
-        .unwrap_err()
-        .unwrap_stdout();
-    let expected_help = "\
-Usage: <USER> <API_KEY>
-
-Available positional items:
-    <USER>     github user
-               in two lines
-    <API_KEY>  api key to use
-
-Available options:
-    -h, --help  Prints help information
-";
-    assert_eq!(expected_help, help);
-}
-
-#[test]
-fn help_for_positional() {
-    let c = positional("C").help("help for\nc");
-    let d = positional("DDD").help("help for\nddd");
-    let parser = construct!(c, d).to_options();
-    let help = parser
-        .run_inner(Args::from(&["--help"]))
-        .unwrap_err()
-        .unwrap_stdout();
-
-    let expected_help = "\
-Usage: <C> <DDD>
-
-Available positional items:
-    <C>    help for
-           c
-    <DDD>  help for
-           ddd
-
-Available options:
-    -h, --help  Prints help information
-";
-    assert_eq!(expected_help, help);
 }
 
 #[test]
@@ -1602,61 +1515,6 @@ fn command_with_req_parameters() {
 }
 
 #[test]
-#[should_panic(expected = "bpaf usage BUG: all positional")]
-fn positional_help_complain_1() {
-    let a = positional("a");
-    let b = short('b').switch();
-    let parser = construct!(a, b).to_options();
-
-    parser
-        .run_inner(Args::from(&["--help"]))
-        .unwrap_err()
-        .unwrap_stderr();
-}
-
-#[test]
-#[should_panic(expected = "bpaf usage BUG: all positional")]
-fn positional_help_complain_2() {
-    let a = positional("a");
-    let b = short('b').switch();
-    let ba = construct!(b, a);
-    let c = short('c').switch();
-    let parser = construct!(ba, c).to_options();
-
-    parser
-        .run_inner(Args::from(&["--help"]))
-        .unwrap_err()
-        .unwrap_stderr();
-}
-
-#[test]
-#[should_panic(expected = "bpaf usage BUG: all positional")]
-fn positional_help_complain_3() {
-    let a = positional("a");
-    let b = short('b').argument("B");
-    let ba = construct!([b, a]);
-    let c = short('c').switch();
-    let parser = construct!(ba, c).to_options();
-
-    parser
-        .run_inner(Args::from(&["--help"]))
-        .unwrap_err()
-        .unwrap_stderr();
-}
-
-#[test]
-fn positional_help_complain_4() {
-    let a = positional("a");
-    let b = short('b').argument("B");
-    let parser = construct!([b, a]).to_options();
-
-    parser
-        .run_inner(Args::from(&["--help"]))
-        .unwrap_err()
-        .unwrap_stdout();
-}
-
-#[test]
 fn suggestion_for_equals_1() {
     let parser = short('p').long("par").argument("P").to_options();
 
@@ -1727,14 +1585,4 @@ fn custom_usage_override() {
         r,
         "Usage: hey [-p]\n\nAvailable options:\n    -p\n    -h, --help  Prints help information\n"
     );
-}
-
-#[test]
-fn strictly_positional() {
-    let parser = positional("A").strict().to_options();
-    let r = parser
-        .run_inner(Args::from(&["a"]))
-        .unwrap_err()
-        .unwrap_stderr();
-    assert_eq!(r, "");
 }
